now tokenization is used for path and parsing is used for templates
