+++ src/compile.spec.ts
@@ -1,8 +1,8 @@
-import { compile, ParsedTokens } from './compile'
+import { compile, TokenizedTemplate } from './compile'
 
 describe('compile()', () => {
-  it('returns ParsedTokens', () => {
-    const compiledTokens: ParsedTokens = compile('Hello {{name}}!')
+  it('returns TokenizedTemplate', () => {
+    const compiledTokens: TokenizedTemplate = compile('Hello {{name}}!')
     expect(compiledTokens).toEqual({
       strings: ['Hello ', '!'],
       refs: [['name']],
