+++ src/parse.spec.ts
@@ -1,154 +1,101 @@
-import { parsePath } from './parse'
+import { parseTemplate } from './parse'
 
-interface ISuccessCases {
-  [input: string]: string[]
-}
+describe('parseTemplate()', () => {
+  it('returns the string intact if no interpolation is found', () => {
+    expect(parseTemplate('Hello world')).toEqual({
+      strings: ['Hello world'],
+      paths: [],
+    })
+  })
 
-function stringArrToString(arr: string[]): string {
-  return '[' + arr.map((s) => `'${s}'`).join(', ') + ']'
-}
+  it('supports customized tags', () => {
+    expect(parseTemplate('Hello {name}!', { tags: ['{', '}'] })).toEqual({
+      strings: ['Hello ', '!'],
+      paths: ['name'],
+    })
+  })
 
-describe('parsePath()', () => {
-  describe('success cases:', () => {
-    // input: expected output
-    const testCases: ISuccessCases = {
-      a: ['a'],
-      ' a': ['a'],
-      '  a': ['a'],
-      'a ': ['a'],
-      'a  ': ['a'],
-      ' a ': ['a'],
-      '  a  ': ['a'],
-      'a.b': ['a', 'b'],
-      'person.name': ['person', 'name'],
-      'a. b ': ['a', 'b'],
-      'a . b ': ['a', 'b'],
-      ' a . b ': ['a', 'b'],
-      'a. b': ['a', 'b'],
-      'a.b ': ['a', 'b'],
-      'a.b.c': ['a', 'b', 'c'],
-      'a["b"]': ['a', 'b'],
-      'a[" b "]': ['a', ' b '],
-      'a[""]': ['a', ''],
-      "a['b']": ['a', 'b'],
-      "a[ 'b' ]": ['a', 'b'],
-      'a["b"].c': ['a', 'b', 'c'],
-      'a[ "b" ].c': ['a', 'b', 'c'],
-      'a [ "b" ] .c': ['a', 'b', 'c'],
-      'a._.c': ['a', '_', 'c'],
-      '.a': ['a'],
-      ' .a': ['a'],
-      '\n.a': ['a'],
-      '\n .a': ['a'],
-      ' \n .a': ['a'],
-      '.\na': ['a'],
-      '.a\n': ['a'],
-      ' . a': ['a'],
-      '. a': ['a'],
-      '_["b"].c': ['_', 'b', 'c'],
-      '__["b"].c': ['__', 'b', 'c'],
-      '_._b.c': ['_', '_b', 'c'],
-      'a.$.c': ['a', '$', 'c'],
-      '': [],
-      ' ': [],
-      '["a"]': ['a'],
-      'a.33': ['a', '33'],
-      'a[0]': ['a', '0'],
-      'a[11]': ['a', '11'],
-      'a[+11]': ['a', '11'],
-      'a[ +12]': ['a', '12'],
-      'a[ + 13]': ['a', '13'],
-      'a[+ 14]': ['a', '14'],
-      'a["15"]': ['a', '15'],
-      'a[ "16"]': ['a', '16'],
-      'a["17" ]': ['a', '17'],
-      'a["b"]["c"]': ['a', 'b', 'c'],
-      '["a"]["b"]["c"]': ['a', 'b', 'c'],
-      '["a"].b["c"]': ['a', 'b', 'c'],
-      'a["b"].c["d"]': ['a', 'b', 'c', 'd'],
-      'a["b"].c["d"].e': ['a', 'b', 'c', 'd', 'e'],
-      'a["+1.1"]': ['a', '+1.1'],
-      '[13]': ['13'],
-      '[17].c': ['17', 'c'],
-      'a["["]': ['a', '['],
-      'a[ "["]': ['a', '['],
-      'a["[" ]': ['a', '['],
-      'a["]"]': ['a', ']'],
-      'a[ "]"]': ['a', ']'],
-      'a["]" ]': ['a', ']'],
-      'a[1234567890123456]': ['a', '1234567890123456'],
-      'a[0001234567890123456]': ['a', '1234567890123456'],
-      'a[001]': ['a', '1'],
-    }
+  it('throws if the open and close tag are the same', () => {
+    expect(() => parseTemplate('Hello |name|!', { tags: ['|', '|'] })).toThrow(TypeError)
+  })
 
-    for (const [input, output] of Object.entries(testCases)) {
-      it(`'${input}'  ⇨  ${stringArrToString(output)}`, () => {
-        expect(parsePath(input)).toEqual(output)
-      })
+  it('throws if the open tag contains the close tag', () => {
+    expect(() => parseTemplate('Hello {{name}!', { tags: ['{{', '{'] })).toThrow(Error)
+  })
+
+  it('throws if the open and close tag are the same', () => {
+    expect(() => parseTemplate('Hello {name}}!', { tags: ['}', '}}'] })).toThrow(Error)
+  })
+
+  it('returns an empty string and no paths when the template is an empty string', () => {
+    expect(parseTemplate('')).toEqual({
+      strings: [''],
+      paths: [],
+    })
+  })
+
+  it('handles interpolation correctly at the start of the template', () => {
+    expect(parseTemplate('{{name}}! How are you?')).toEqual({
+      strings: ['', '! How are you?'],
+      paths: ['name'],
+    })
+  })
+
+  it('handles interpolation correctly at the end of the template', () => {
+    expect(parseTemplate('My name is {{name}}')).toEqual({
+      strings: ['My name is ', ''],
+      paths: ['name'],
+    })
+  })
+
+  it('trims path', () => {
+    const { paths } = parseTemplate('My name is {{  name  }}')
+    if (paths.length) {
+      expect(paths[0]).toBe('name')
     }
   })
 
-  describe('SyntaxError cases:', () => {
-    // all these strings throw a syntax error
-    const syntaxErrorsCases: string[] = [
-      'a.',
-      'a..',
-      'a..b',
-      'a ..b',
-      'a . .b',
-      'a . . b',
-      'a .. b',
-      '..',
-      '. .',
-      ' . . ',
-      ' .. ',
-      ' .. . ',
-      ' ... ',
-      '.["a"]',
-      '.',
-      ' . ',
-      'a["b"]c',
-      'a.["b"]c',
-      'a.["b"]c["d"]',
-      'a["b"]c["d"]',
-      'a["b"].c["d"]e',
-      'a["b"].c.',
-      'a ["b"] c',
-      'a["]',
-      'a.[b]',
-      'a[\'b"]',
-      'a[',
-      'a]',
-      'a]]',
-      'a[[',
-      'a["b\']',
-      'a["b`]',
-      'a[11"]',
-      'a[`11]',
-      'a[ `11 ]',
-      'a[[]]',
-      'a[-11]',
-      'a[b]',
-      'a[11x]',
-      'a[1.1]',
-      'a[+1.1]',
-      'a[-1.1]',
-      '.["a"]',
-      'name[a].',
-      'name,',
-      'name;',
-      'a[12345678901234567]',
-    ]
-
-    for (const input of syntaxErrorsCases) {
-      it(`throws SyntaxError for "${input}"`, () => {
-        expect(() => parsePath(input)).toThrow(SyntaxError)
-      })
-    }
-
-    it('throws type error for invalid input types', () => {
-      expect(() => parsePath((undefined as unknown) as string)).toThrow(TypeError)
-      expect(() => parsePath((13 as unknown) as string)).toThrow(TypeError)
+  it('can handle a close tag without an open tag', () => {
+    expect(parseTemplate('Hi}} {{name}}')).toEqual({
+      strings: ['Hi}} ', ''],
+      paths: ['name'],
     })
+    expect(parseTemplate('Hi {{name}} }}')).toEqual({
+      strings: ['Hi ', ' }}'],
+      paths: ['name'],
+    })
+  })
+
+  it('throws a syntax error if the open tag is not closed', () => {
+    expect(() => parseTemplate('Hi {{')).toThrow(
+      new SyntaxError(
+        'Missing "}}" in the template for the "{{" at position 3 within 1000 characters'
+      )
+    )
+  })
+
+  it('does not throw an error if there is a close tag without an open tag', () => {
+    expect(() => parseTemplate('Hi}} ')).not.toThrow()
+  })
+
+  it('throws a syntax error if the path is an empty string', () => {
+    expect(() => parseTemplate('Hi {{}}')).toThrow(
+      new SyntaxError('Unexpected "}}" tag found at position 3')
+    )
+  })
+
+  it('throws a syntax error if the value name is just spaces', () => {
+    expect(() => parseTemplate('Hi {{ }}')).toThrow(
+      new SyntaxError('Unexpected "}}" tag found at position 3')
+    )
+  })
+
+  it('throws for nested open and close tag', () => {
+    expect(() => parseTemplate('Hello {{ {{name}} }}!')).toThrow()
+  })
+
+  it('throws if the path is too long', () => {
+    expect(() => parseTemplate('Hej {{n2345}}!', { maxPathLen: 5 })).not.toThrow()
+    expect(() => parseTemplate('Hej {{n2345}}!', { maxPathLen: 4 })).toThrow()
   })
 })
